:recycle: Refactor TransactionList filenames to more representative ones
