+++ src/modules/Transactions/components/TransactionList/TransactionList.js
@@ -15,10 +15,10 @@ import ResultScreen from 'components/screens/ResultScreen';
 import EmptyIllustrationSvg from 'assets/svgs/EmptyIllustrationSvg';
 import ErrorIllustrationSvg from 'assets/svgs/ErrorIllustrationSvg';
 import CaretSvg from 'assets/svgs/CaretSvg';
-import TransactionRow from '../TransactionRow';
+import TransactionRow from '../TransactionRow/TransactionRow';
 
-import getTransactionListStyles from './styles';
-import { NO_OF_TRANSACTIONS_ON_OVERVIEW } from './constants';
+import getTransactionListStyles from './TransactionList.styles';
+import { NO_OF_TRANSACTIONS_ON_OVERVIEW } from './TransactionList.constants';
 
 export default function TransactionList({ mode = 'overview', style }) {
   const navigation = useNavigation();
