+++ src/modules/Transactions/components/TransactionList/components/TransactionListSkeleton.js
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native';
+import React from 'react';
+
+export default function TransactionListSkeleton() {
+  return (
+    <View>
+      <Text>TransactionListSkeleton</Text>
+    </View>
+  );
+}
