+++ src/modules/Transactions/components/TransactionRow/components/TransactionAmount.js
@@ -4,10 +4,7 @@ import { useSelector } from 'react-redux';
 
 import { fromRawLsk } from 'utilities/conversions';
 import FormattedNumber from 'components/shared/formattedNumber';
-import CircleCrossedSvg from 'assets/svgs/CircleCrossedSvg';
-import CheckSvg from 'assets/svgs/CheckSvg';
-import SandClockSvg from 'assets/svgs/SandClockSvg';
-import { useTransactionAssets } from '../../hooks/useTransactionAssets';
+import { useTransactionAssets } from '../../../hooks/useTransactionAssets';
 
 export function TransactionAmount({ transaction, style }) {
   const transactionAssets = useTransactionAssets({ transaction });
@@ -29,31 +26,6 @@ export function TransactionAmount({ transaction, style }) {
         </FormattedNumber>
       </Text>
     );
-  }
-
-  return children;
-}
-
-export function TransactionStatus({ transaction }) {
-  let children = null;
-
-  const props = { height: 14, width: 14 };
-
-  switch (transaction.executionStatus) {
-    case 'success':
-      children = <CheckSvg {...props} />;
-      break;
-
-    case 'pending':
-      children = <SandClockSvg {...props} />;
-      break;
-
-    case 'fail':
-      children = <CircleCrossedSvg {...props} />;
-      break;
-
-    default:
-      break;
   }
 
   return children;
